2.2124
* Zip64 Support
* Reduced memory consumption on OSX/Linux when downloading files
* Reliability improvements for
  QueryCurrentModUpdate/Modio::ModProgressInfo
* Warning cleanup esp for non-MSVC compilers

+++ modio/modio/file/ModioFile.h
@@ -57,7 +57,7 @@ namespace Modio
 				  FilePath(std::move(Other.FilePath)),
 				  FileStrand(std::move(Other.FileStrand)) {};
 
-			std::size_t GetFileSize()
+			std::uint64_t GetFileSize()
 			{
 				return get_implementation()->GetSize();
 			}
